# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: harness-telemetry API
  version: 0.1.0
servers:
  - url: /api
paths:
...
  /analytics/dataset:
    get:
      operationId: getAnalyticsDataset
      summary: The live analytics dataset for the filter
      description: >-
        Sessions with a hook event inside the window and the aggregates, cards and summary built
        over them. Every signed-in user sees every user; user is only a filter.
      parameters:
        - name: days
          in: query
          required: false
          description: The window is the last days days; without it the default window
          schema:
            type: integer
            minimum: 1
            maximum: 365
        - $ref: "#/components/parameters/AnalyticsAgent"
        - name: project
          in: query
          required: false
          schema:
            type: string
        - $ref: "#/components/parameters/AnalyticsUser"
        - name: tz
          in: query
          required: false
          description: >-
            The browser's zone, an IANA name or an offset ±HH:MM; each session's daily is laid out
            by its days. UTC days without it.
          schema:
            type: string
            maxLength: 64
        - name: kind
          in: query
          required: false
          style: form
          explode: false
          schema:
            type: array
            items:
              $ref: "#/components/schemas/AnalyticsSessionKind"
        - name: If-None-Match
          in: header
          required: false
          description: The ETag of a dataset the client holds; the same dataset answers 304
          schema:
            type: string
      responses:
        "200":
          description: The dataset
          headers:
            ETag:
              description: Identifies this dataset's body; the same body has the same ETag
              schema:
                type: string
            Cache-Control:
              description: private, no-cache — the browser keeps the body and asks again each time
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsDataset"
        "304":
          description: The dataset If-None-Match names is still the current one
          headers:
            ETag:
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
        "400":
          description: invalid_filter — a parameter is out of range; message names it
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/Error"
  /analytics/team:
    get:
      operationId: getAnalyticsTeam
      summary: Every person's numbers for «Команда»
      description: >-
        One row per person with sessions in the period, counted as «Обзор» counts one person's
        sessions; no sessions and no aggregates. Built from the dataset of everyone for the period,
        which the server caches; agent, project and kind only choose its sessions.
      parameters:
        - name: days
          in: query
          required: false
          description: The period is the last days days; 365 without it
          schema:
            type: integer
            minimum: 1
            maximum: 365
        - $ref: "#/components/parameters/AnalyticsAgent"
        - name: project
          in: query
          required: false
          description: A project's name; «без проекта» chooses the sessions without one
          schema:
            type: string
        - name: kind
          in: query
          required: false
          description: work (the default) leaves the service sessions out, all keeps them
          schema:
            type: string
            enum: [work, all]
      responses:
        "200":
          description: The people and what the filters of the page need
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsTeam"
        "400":
          description: invalid_filter — a parameter is out of range; message names it
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          $ref: "#/components/responses/Error"
  /analytics/sessions/{id}:
    get:
      operationId: getAnalyticsSession
      summary: The timeline of one session
      description: >-
        id is the agent's session id, not ours; agent and user tell apart sessions that share it.
        not_found — no such session; ambiguous_session (409) — several sessions match, pass agent and user.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/AnalyticsAgent"
        - $ref: "#/components/parameters/AnalyticsUser"
      responses:
        "200":
          description: The timeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsTimeline"
        default:
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    AnalyticsAgent:
      name: agent
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/AnalyticsAgent"
    AnalyticsUser:
      name: user
      in: query
      required: false
      description: A filter by user, not an access check; every signed-in user sees every user
      schema:
        type: string
        format: uuid
...
  responses:
...
    Error:
      description: Any 4xx or 5xx answer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      description: The common body of every 4xx and 5xx response
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code, e.g. not_found
        message:
          type: string
          description: Human-readable description
    AnalyticsAgent:
      type: string
      enum: [claude, codex]
    AnalyticsSessionKind:
      type: string
      description: >-
        system — a service task of Codex; agent — a session another agent started (a codex exec
        run, a first prompt that is a cross-session message or an orchestrator's brief);
        automation — a <heartbeat> prompt; user — the rest
      enum: [system, agent, automation, user]
    AnalyticsSeverity:
      type: string
      enum: [bad, warn, info]
    AnalyticsDataset:
      description: >-
        The live dataset: the colleague's dataset.json in our form. Unknown is null, never 0.
        Aggregates are built over the filtered sessions; by_session lets the page recount a subset.
      type: object
      required:
        - schema_version
        - variant
        - generated_at
        - window
        - filter
        - facets
        - summary
        - pricing
        - gaps
        - sessions
        - tools
        - mcp
        - commands
        - permissions
        - friction
        - skills
        - findings
        - done
        - checks_catalog
      properties:
        schema_version:
          type: integer
        variant:
          type: string
          enum: [live]
        generated_at:
          type: string
          format: date-time
        window:
          $ref: "#/components/schemas/AnalyticsWindow"
        filter:
          $ref: "#/components/schemas/AnalyticsFilter"
        facets:
          $ref: "#/components/schemas/AnalyticsFacets"
        summary:
          $ref: "#/components/schemas/AnalyticsSummary"
        pricing:
          $ref: "#/components/schemas/AnalyticsPricing"
        gaps:
          type: array
          description: What is not collected and why
          items:
            type: string
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsSession"
        tools:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsToolRow"
        mcp:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsMcpRow"
        commands:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsCommandRow"
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsPermissionRow"
        friction:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsFriction"
        skills:
          $ref: "#/components/schemas/AnalyticsSkills"
        findings:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsFinding"
        done:
          type: array
          description: Applied changes with their effect; empty in the live variant
          items:
            type: object
            additionalProperties: true
        checks_catalog:
          type: array
          description: The catalogue of checks; empty in the live variant
          items:
            $ref: "#/components/schemas/AnalyticsCheckDef"
    AnalyticsTeam:
      description: >-
        «Команда» for the period: each person's numbers over their chosen sessions. Unknown is null,
        never 0.
      type: object
      required: [generated_at, window, people, projects, has_system, gaps, session_gaps]
      properties:
        generated_at:
          type: string
          format: date-time
        window:
          description: The period; for «Всё», the window of the dataset
          allOf:
            - $ref: "#/components/schemas/AnalyticsWindow"
        people:
          description: The people with a chosen session, in no order
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsTeamPerson"
        projects:
          description: >-
            The projects of every session of the dataset (twice the period), whatever agent,
            project and kind say, for the filter of projects; «без проекта» stands for the
            sessions without one
          type: array
          items:
            type: string
        has_system:
          description: A session of the dataset is a service one, so the filter of them matters
          type: boolean
        gaps:
          description: The dataset's notes that are not about one session
          type: array
          items:
            type: string
        session_gaps:
          description: How many of the dataset's notes are about one session; they stay in its timeline
          type: integer
    AnalyticsTeamPerson:
      type: object
      required:
        - user_id
        - user_name
        - sessions
        - user_min
        - agent_min
        - cost_usd
        - cost_partial
        - error_rate
        - friction_sessions
        - last_active
        - pulse
      properties:
        user_id:
          type: string
          format: uuid
        user_name:
          type: string
        sessions:
          description: The chosen sessions of kind user
          type: integer
        user_min:
          description: The person's minutes over the sessions that are neither automation nor agent
          type: number
          format: double
          nullable: true
        agent_min:
          description: The agent's active minutes over the same sessions
          type: number
          format: double
          nullable: true
        cost_usd:
          description: The cost of the chosen sessions with a known cost; null when none has one
          type: number
          format: double
          nullable: true
        cost_partial:
          description: Some chosen session has no cost, so cost_usd is a lower bound
          type: boolean
        error_rate:
          description: Tool errors over the calls with a known outcome; null without such calls
          type: number
          format: double
          nullable: true
        friction_sessions:
          description: The chosen sessions with a friction signal
          type: integer
        last_active:
          description: The latest end, or start without one, of the chosen sessions
          type: string
          format: date-time
        pulse:
          description: >-
            The person's hook events by hour over the 5 whole UTC days before today, whatever the
            filter; null when the telemetry store could not be read
          allOf:
            - $ref: "#/components/schemas/AnalyticsTeamPulse"
          nullable: true
...
    AnalyticsWindow:
      type: object
      required: [from, to]
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
    AnalyticsFilter:
      description: The filter the dataset was built for, with its defaults filled in
      type: object
      required: [days, kind]
      properties:
        days:
          type: integer
        agent:
          $ref: "#/components/schemas/AnalyticsAgent"
        project:
          type: string
        user:
          type: string
          format: uuid
        kind:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsSessionKind"
    AnalyticsFacets:
      description: The values the filter can take in the window
      type: object
      required: [agents, projects, users, kinds]
      properties:
        agents:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsAgent"
        projects:
          type: array
          items:
            type: string
        users:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsUserRef"
        kinds:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsSessionKind"
    AnalyticsUserRef:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    AnalyticsSummary:
      type: object
      required: [coverage, funnel, period]
      properties:
        coverage:
          $ref: "#/components/schemas/AnalyticsCoverage"
        funnel:
          $ref: "#/components/schemas/AnalyticsFunnel"
        period:
          $ref: "#/components/schemas/AnalyticsPeriodTotals"
...
    AnalyticsFunnel:
      description: The first three steps of the improvement cycle
      type: object
      required: [sessions, signals, topics]
      properties:
        sessions:
          type: integer
          description: Sessions of kind user and automation
        signals:
          type: integer
          description: Friction episodes
        topics:
          type: integer
          description: Cards whose scope is not collection
...
    AnalyticsPricing:
      description: An estimate, not a bill
      type: object
      required: [basis, note, models]
      properties:
        basis:
          type: string
        note:
          type: string
        models:
          type: object
          description: USD per million tokens by model
          additionalProperties:
            $ref: "#/components/schemas/AnalyticsModelPrice"
...
    AnalyticsSession:
      description: One session; the key is (user_id, agent, id)
      type: object
      required:
        - id
        - short
        - agent
        - user_id
        - user_name
        - host
        - records
        - project
        - cwd
        - branch
        - start
        - end
        - wall_min
        - active_min
        - user_min
        - turns
        - prompts
        - corrections
        - reqs
        - model
        - cache_hit
        - cost_usd
        - cost_basis
        - calls
        - errors
        - unknown_results
        - outcome_known
        - commits
        - prs
        - added
        - removed
        - first
        - title
        - flags
        - compactions
        - aborted
        - waits
        - wait_min
        - kind
        - kind_reason
        - sources
        - daily
      properties:
        id:
          type: string
          description: The agent's session id
        short:
          type: string
          description: The first 8 and the last 4 characters of id
        agent:
          $ref: "#/components/schemas/AnalyticsAgent"
        user_id:
          type: string
          format: uuid
        user_name:
          type: string
        host:
          type: string
        records:
          type: integer
          description: Hook events of the session
        project:
          type: string
        cwd:
          type: string
        branch:
          type: string
          nullable: true
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        wall_min:
          type: number
        active_min:
          type: number
        user_min:
          type: number
          nullable: true
        turns:
          type: integer
        prompts:
          type: integer
        corrections:
          type: integer
          nullable: true
        reqs:
          type: integer
          nullable: true
        model:
          type: string
          nullable: true
        tok:
          allOf:
            - $ref: "#/components/schemas/AnalyticsTokens"
          nullable: true
        cache_hit:
          type: number
          nullable: true
        cost_usd:
          type: number
          nullable: true
        cost_basis:
          type: string
          nullable: true
        calls:
          type: integer
        errors:
          type: integer
        unknown_results:
          type: integer
        outcome_known:
          type: integer
        commits:
          type: integer
        prs:
          type: integer
        added:
          type: integer
          nullable: true
        removed:
          type: integer
          nullable: true
        first:
          type: string
          description: The first prompt, up to 200 characters
        title:
          type: string
        flags:
          type: array
          items:
            type: string
        compactions:
          type: integer
        aborted:
          type: integer
        waits:
          type: integer
        wait_min:
          type: number
          nullable: true
        kind:
          $ref: "#/components/schemas/AnalyticsSessionKind"
        kind_reason:
          type: string
          nullable: true
        sources:
          $ref: "#/components/schemas/AnalyticsSources"
        daily:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsDaily"
...
    AnalyticsToolRow:
      type: object
      required: [name, kind, server, calls, errors, unknown, p50_s, p95_s, by_session]
      properties:
        name:
          type: string
        kind:
          type: string
          enum: [builtin, mcp, nested]
        server:
          type: string
          nullable: true
        calls:
          type: integer
        errors:
          type: integer
        unknown:
          type: integer
        p50_s:
          type: number
          nullable: true
        p95_s:
          type: number
          nullable: true
        by_session:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/AnalyticsCallCounts"
    AnalyticsMcpRow:
      type: object
      required: [server, calls, errors, unknown, avg_s, by_session]
      properties:
        server:
          type: string
        calls:
          type: integer
        errors:
          type: integer
        unknown:
          type: integer
        avg_s:
          type: number
          nullable: true
        by_session:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/AnalyticsCallCounts"
    AnalyticsCommandRow:
      type: object
      required: [cmd, project, runs, failed, unknown, p95_s, total_min, by_session]
      properties:
        cmd:
          type: string
        project:
          type: string
        runs:
          type: integer
        failed:
          type: integer
        unknown:
          type: integer
        p95_s:
          type: number
          nullable: true
        total_min:
          type: number
        by_session:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/AnalyticsRunCounts"
    AnalyticsPermissionRow:
      type: object
      required: [label, count, by_session]
      properties:
        label:
          type: string
        count:
          type: integer
        by_session:
          type: object
          additionalProperties:
            type: integer
...
    AnalyticsFriction:
      type: object
      required: [key, name, how, sev, sessions, by_session, count, cost_usd, evidence]
      properties:
        key:
          type: string
          description: retry, thrash, mcpfail, coldcache, compact, reread, wait, abort
        name:
          type: string
        how:
          type: string
        sev:
          $ref: "#/components/schemas/AnalyticsSeverity"
        sessions:
          type: array
          items:
            type: string
        by_session:
          type: object
          description: Episodes by session
          additionalProperties:
            type: integer
        count:
          type: integer
          nullable: true
        cost_usd:
          type: number
          nullable: true
          description: The cost of the episodes themselves; only signals with an attributable cost
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsEvidence"
    AnalyticsSkills:
      type: object
      required: [available, rows]
      properties:
        available:
          type: integer
          nullable: true
          description: null until the snapshot of available skills is collected
        rows:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsSkillRow"
        gantt:
          $ref: "#/components/schemas/AnalyticsGantt"
...
    AnalyticsFinding:
      description: A «What to fix» card; Deep fields are added to this schema by stage 4
      type: object
      required:
        - id
        - sev
        - title
        - what
        - ev
        - impact
        - impact_by_session
        - kind
        - scope
        - readiness
        - decision
        - execution
        - effect
        - source
        - lang
        - sessions
        - episodes
        - cause
        - alternative_causes
        - preconditions
        - exceptions
        - verification
        - rollback
        - expected_effect
      properties:
        id:
          type: string
        sev:
          $ref: "#/components/schemas/AnalyticsSeverity"
        title:
          type: string
        what:
          type: string
        ev:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsEvidence"
        impact:
          description: null when nothing supports it
          allOf:
            - $ref: "#/components/schemas/AnalyticsImpact"
          nullable: true
        impact_by_session:
          type: object
          additionalProperties:
            type: number
        impact_unit:
          type: string
        where:
          type: string
        snip:
          type: string
        kind:
          type: string
          enum: [personalization, project_rule, skill, hook, script, automation, diagnostic, habit]
        pattern_id:
          type: string
        scope:
          type: string
          enum: [work, collection]
        readiness:
          type: string
          enum: [hypothesis, needs_spec, prepared]
        decision:
          $ref: "#/components/schemas/ProposalDecisionStatus"
        execution:
          $ref: "#/components/schemas/ProposalExecutionStatus"
        effect:
          $ref: "#/components/schemas/ProposalEffectStatus"
        source:
          type: string
        lang:
          type: string
        sessions:
          type: array
          items:
            type: string
        episodes:
          type: integer
        cause:
          type: string
          nullable: true
        alternative_causes:
          type: array
          items:
            type: string
        preconditions:
          type: array
          items:
            type: string
        exceptions:
          type: array
          items:
            type: string
        verification:
          type: string
          nullable: true
        rollback:
          type: string
          nullable: true
        expected_effect:
          type: string
          nullable: true
    AnalyticsCheckDef:
      type: object
      required: [id, name, how]
      properties:
        id:
          type: string
        name:
          type: string
        how:
          type: string
    AnalyticsTimeline:
      description: The timeline of one session, the colleague's sessions/<id>.json
      type: object
      required: [id, events, cum, ctx, tools, turns]
      properties:
        id:
          type: string
        events:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsEvent"
        cum:
          type: array
          description: Minutes from the start → cost so far in USD (estimate)
          items:
            $ref: "#/components/schemas/AnalyticsPoint"
        ctx:
          type: array
          description: Minutes from the start → the request's input in thousands of tokens
          items:
            $ref: "#/components/schemas/AnalyticsPoint"
        tools:
          type: object
          description: Calls by tool, the commonest first
          additionalProperties:
            type: integer
        turns:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsTurn"
    AnalyticsPoint:
      type: array
      minItems: 2
      maxItems: 2
      items:
        type: number
    AnalyticsEvent:
      type: object
      required: [line, at, turn, k, x, note, side, tool]
      properties:
        line:
          type: integer
        at:
          type: string
          format: date-time
        turn:
          type: integer
        k:
          type: string
          enum: [prompt, answer, tool, err, api, wait, compact, abort, skill, agent]
        x:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        side:
          type: string
          nullable: true
        tool:
          type: string
          nullable: true
        src_line:
          type: integer
          nullable: true
          description: >-
            The line of the session's transcript the event is built from, counted from 1 as a
            Deep review's L<n> counts it: a call by its tool_use id or call_id, a prompt by its
            promptId or its Codex turn's start, a Codex answer by its turn's end. Null when the
            transcript names none.
        src_kind:
          type: string
          nullable: true
          description: >-
            The kind of the transcript file src_line counts in, main or subagent; null with
            src_line.
    AnalyticsTurn:
      type: object
      required: [turn, turn_id, start, end, dur_ms, state, prompt_line, prompt, answer, tok]
      properties:
        turn:
          type: integer
        turn_id:
          type: string
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
          nullable: true
        dur_ms:
          type: integer
          format: int64
        state:
          type: string
        prompt_line:
          type: integer
          nullable: true
        prompt:
          type: string
          nullable: true
        answer:
          type: string
          nullable: true
        tok:
          description: null when the session has no model request
          allOf:
            - $ref: "#/components/schemas/AnalyticsTurnTokens"
          nullable: true
...